rename package to com.auth0.android.auth0
