+++ app/src/main/AndroidManifest.xml
@@ -65,7 +65,7 @@
 
         <!--Auth0 Lock Embedded WebView-->
         <activity
-            android:name="com.auth0.oauth2.WebViewActivity"
+            android:name="com.auth0.android.auth0.WebViewActivity"
             android:theme="@style/MyLock.Theme" />
         <!--Auth0 Lock Embedded WebView End-->
