+++ app/src/main/java/com/auth0/android/lock/app/AuthProviderHandler.java
@@ -29,8 +29,8 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import com.auth0.android.lock.provider.AuthProviderResolver;
-import com.auth0.oauth2.AuthCallback;
-import com.auth0.oauth2.AuthProvider;
+import com.auth0.android.auth0.AuthCallback;
+import com.auth0.android.auth0.AuthProvider;
 
 public class AuthProviderHandler implements AuthProviderResolver {
