+++ auth0/src/main/java/com/auth0/android/auth0/CallbackHelper.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.oauth2;
+package com.auth0.android.auth0;
 
 import android.net.Uri;
 import android.support.annotation.NonNull;
