+++ auth0/src/main/java/com/auth0/android/auth0/OAuth2WebAuthProvider.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.oauth2;
+package com.auth0.android.auth0;
 
 import android.app.Activity;
 import android.content.Intent;
