+++ auth0/src/test/java/com/auth0/android/auth0/AuthProviderTest.java
@@ -22,16 +22,11 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.oauth2;
+package com.auth0.android.auth0;
 
 import android.app.Activity;
 import android.support.annotation.NonNull;
 import android.support.v4.content.PermissionChecker;
-
-import com.auth0.oauth2.AuthCallback;
-import com.auth0.oauth2.AuthProvider;
-import com.auth0.oauth2.AuthorizeResult;
-import com.auth0.oauth2.PermissionHandler;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -48,7 +43,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
 @RunWith(RobolectricGradleTestRunner.class)
-@Config(constants = com.auth0.oauth2.BuildConfig.class, sdk = 21, manifest = Config.NONE)
+@Config(constants = com.auth0.android.auth0.BuildConfig.class, sdk = 21, manifest = Config.NONE)
 public class AuthProviderTest {
 
     @Mock
