+++ lib/src/main/java/com/auth0/android/lock/PasswordlessLockActivity.java
@@ -69,11 +69,11 @@ import com.auth0.android.lock.views.PasswordlessLockView;
 import com.auth0.authentication.AuthenticationAPIClient;
 import com.auth0.authentication.result.Credentials;
 import com.auth0.callback.BaseCallback;
-import com.auth0.oauth2.AuthCallback;
-import com.auth0.oauth2.AuthProvider;
-import com.auth0.oauth2.AuthorizeResult;
-import com.auth0.oauth2.CallbackHelper;
-import com.auth0.oauth2.OAuth2WebAuthProvider;
+import com.auth0.android.auth0.AuthCallback;
+import com.auth0.android.auth0.AuthProvider;
+import com.auth0.android.auth0.AuthorizeResult;
+import com.auth0.android.auth0.CallbackHelper;
+import com.auth0.android.auth0.OAuth2WebAuthProvider;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.otto.Bus;
 import com.squareup.otto.Subscribe;
